\name{qDEA-package}
\alias{qDEA-package}
\alias{qDEA}
\docType{package}
\title{
qDEA bootstrapping using fortran interface to glpk
}
\description{
qDEA bootstrpping using fortran interface to glpk
}
\details{
\tabular{ll}{
Package: \tab qDEA\cr
Type: \tab Package\cr
Version: \tab 0.2\cr
Date: \tab 2016-09-15\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}



}
\author{
Joe Atwood

Maintainer: Joe Atwood <jatwood@montana.edu>
}

\examples{
#not run
}
